REFACTOR : 예약 신청 uri 변경
