+++ src/main/java/com/meta/overwash/service/ReservationService.java
@@ -6,13 +6,24 @@ import com.meta.overwash.domain.ReservationDTO;
 import com.meta.overwash.domain.WashingCompleteDTO;
 
 public interface ReservationService {
-	public List<ReservationDTO> getList();
-	public List<ReservationDTO> getListEach(Long memberId);
-	public ReservationDTO getListEachOne(ReservationDTO reservation);
-	public List<ReservationDTO> getListMember(Long memberId);
-	public void register(ReservationDTO reservation);
-	public void updateReservationStatus(ReservationDTO reservation);
-	public void registerWashingComplete(WashingCompleteDTO washComplete);
-	public Long getMemberId(String username);
 	
+	public List<ReservationDTO> getList();
+
+	public List<ReservationDTO> getListEach(Long memberId);
+
+	public ReservationDTO getListEachOne(ReservationDTO reservation);
+
+	public List<ReservationDTO> getListMember(Long memberId);
+
+	// 예약 신청
+	public void insert(ReservationDTO reservation);
+
+	// 예약 상태 변경
+	public void updateReservationStatus(ReservationDTO reservation);
+
+	// 세탁 완료에 추가 ?
+	public void insertWashingComplete(WashingCompleteDTO washComplete);
+
+	public Long getMemberId(String username);
+
 }
